Remove XML file storage and serializers, add IRC server models and IrcServerStorage interface
Deleted XML file storage and related serializers since other alternatives are used for data storage. Added IrcChannel, IrcUser, IrcUserChannel models and IIrcServerStorage interface to fulfil the requirements for an IRC server. This change simplifies the system architecture and promotes the usage of better data storage solutions. It also prepares the system for an IRC server implementation.