Remove unnecessary gaming core libraries
Deleted the libraries including AssetProvider, File providers, ObjParsing, Polygon parsers, and others because they are not required in the system anymore. This commit helps in tidying up the project structure and reduces complexity. The decision was taken based on the team discussion to simplify the project structure, as these files and game assets were previously used for test purposes and are not a part of the actual implementation.