Add PackageReadmeFile element to Directory.Build.props
Added a PackageReadmeFile element to the Directory.Build.props file to ensure the readme.md file is included in the generated NuGet package. Also modified the packaging path for the readme.md to align with the new PackageReadmeFile specification. This change will help package users to better understand the package functionality.